+++ electron/preload.js
@@ -8,6 +8,7 @@ contextBridge.exposeInMainWorld('electronAPI', {
   // Node CRUD
   getNodes: params => ipcRenderer.invoke(C.DB_GET_NODES, params),
   getNode: id => ipcRenderer.invoke(C.DB_GET_NODE, id),
+  getNodeNotes: id => ipcRenderer.invoke(C.DB_GET_NODE_NOTES, id),
   createNode: data => ipcRenderer.invoke(C.DB_CREATE_NODE, data),
   updateNode: (id, data) => ipcRenderer.invoke(C.DB_UPDATE_NODE, id, data),
   deleteNode: (id, hard) => ipcRenderer.invoke(C.DB_DELETE_NODE, id, hard),